// frontend/src/types/Filters.tsx
export interface Filters {
    city: string;
    state: string;
    minPrice: string;
    maxPrice: string;
    rooms: string;
    bathrooms: string;
    garage: string;
    squareFeet: string;
    level: string;
    numberOfKitchen: string;
    numberOfBalconies: string;
    hasGarden: string;
    forRent: string;
    yearOfConstruction: string;
    furnished: string;
}